separation of server and client components
