+++ src/app/(protected)/checkout/shipping/page.tsx
@@ -294,12 +294,11 @@ function CheckoutShippingContent() {
           calculatedTotals,
           checkoutSession,
           shippingAddressId: selectedAddressId,
-          billingAddressId: selectedAddressId, // Assuming same as shipping
+          billingAddressId: selectedAddressId,
           urgencyLevel,
           customerId: userProfile?.id,
-          // Explicitly include cart items for order creation
-          useCartItems: true, // This tells the API to use items from the checkout session
-          items: checkoutSession?.items || [], // Include the items for reference
+          useCartItems: true,
+          items: checkoutSession?.items || [],
         };
 
         // Store all checkout data for payment page
